pyproject.toml
[project]
name = "lan-control-plane"
version = "0.1.0"
description = "Self-hosted LAN control plane with WebSocket agents and clients"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "websockets>=13.0",
  "argon2-cffi>=23.1.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.0",
  "mypy>=1.11.0",
  "ruff>=0.6.0",
  "pre-commit>=3.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
pythonpath = [
  "server/src",
  "agent/src",
  "shared/src",
]
testpaths = [
  "server/tests",
  "agent/tests",
  "shared/tests",
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
mypy_path = [
  "server/src",
  "agent/src",
  "shared/src",
]

[[tool.mypy.overrides]]
module = ["uvicorn.*"]
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = [
    "server/src/lan_control_plane_server",
    "agent/src/lan_control_plane_agent",
    "shared/src/lan_control_plane_shared",
]

[tool.hatch.build.targets.sdist]
include = [
    "server/src/lan_control_plane_server",
    "agent/src/lan_control_plane_agent",
    "shared/src/lan_control_plane_shared",
    "server/tests",
    "agent/tests",
    "shared/tests",
    "alembic",
    "README.md",
    "pyproject.toml",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
